narrowing fixes: 5 MWERKS-narrowing sentinels made explicit
Fixes surfacing when the parent sunbright build is compiled without
-Wno-narrowing on sms-native. Each fix preserves the exact wire value
MWERKS produced when it silently accepted the narrowing.

- src/Enemy/emario.cpp:216 — case 0x80000001 in a switch(s32) needs
  an explicit s32 cast (2147483649u → -2147483647 signed).
- src/Enemy/hamukuri.cpp:473-476 — GXColorS10 channel initializers
  0xFFEC/0xFFA6/0xFF74/0xFFD8/0xFFBA/0xFFCE are s16 negative values;
  explicit (s16) casts preserve bit pattern.
- src/JSystem/JAudio/JASystem/JASPlayer_impl.cpp:105,108 — s16 vib/tre
  table sentinels 0xC000 / 0x8001 explicit-cast to (s16).
- src/MSound/MSound.cpp:35 — u16[32] sentinel initializer -1 replaced
  with 0xFFFF (the same wire value, GCC-valid).

Not fixed in this commit (parent's sms-native target retains
-Wno-narrowing to cover the remaining ~30 MWERKS-narrowing sites in
this decomp library; a proper eradication needs per-site type-correcting
across ~500 TUs and is out of scope for a first-party build-hygiene
sweep).

+++ src/Enemy/emario.cpp
@@ -213,7 +213,7 @@ void TEMario::perform(u32 cue, JDrama::TGraphics* graphics)
 
 	for (s32 i = 0; i < mColCount; ++i) {
 		switch (mCollisions[i]->mActorType) {
-		case 0x80000001: {
+		case (s32)0x80000001: {
 			const f32 d = vecDist(mPosition, mCollisions[i]->getPosition());
 			if (d < mEnemyMario->unk42B0) {
 				mCollisions[i]->receiveMessage(this, HIT_MESSAGE_ATTACK);
